Extracted get_current_timestamp into time utils
