+++ include/Utils/TimeUtils.hpp
@@ -1,0 +1,13 @@
+#pragma once
+
+#include <string>
+#include <cstdint>
+#include <chrono>
+#include <iomanip>
+
+/**
+ * @brief Gets the current timestamp as a string.
+ *
+ * @return The current timestamp.
+ */
+std::string get_current_timestamp();
